% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{averageTimeSeries_fast}
\alias{averageTimeSeries_fast}
\title{Average Univariate Time Series}
\usage{
averageTimeSeries_fast(tsList, weights)
}
\arguments{
\item{tsList}{A list of NumericVectors all having the same length.}

\item{weights}{A vector of weights, having the same length as the list.}
}
\value{
The weighted average of the series from the list.
}
\description{
Multiplies a list of univariate time series with a weight vector and
sums up the result to get a weighted average.
}
