% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{averageTimeSeriesList}
\alias{averageTimeSeriesList}
\title{Average Time Series}
\usage{
averageTimeSeriesList(tsList, dissimilarities,
  dissWeightConversion = "minmax")
}
\arguments{
\item{tsList}{A list of numeric vectors/matrices (uni-/multivariate time series).}

\item{dissimilarities}{A numeric vector of dissimilarity values (same length
as the list of time series).}

\item{dissWeightConversion}{Strategy for converting similarities to
dissimilarities. Might be "minmax" (linear), "gaussian" or "laplacian".}
}
\description{
Averages a list of equal-length uni- or multivariate time series based on
dissimilarities (which are normalized, transformed to similarities and used
as weights), puting more weight on series which have a higher similarity
(lower dissimilarity).
}
