% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{averageTimeSeriesMult_fast}
\alias{averageTimeSeriesMult_fast}
\title{Average Multivariate Time Series}
\usage{
averageTimeSeriesMult_fast(tsList, weights)
}
\arguments{
\item{tsList}{A list of NumericMatrix all having the same length.}

\item{weights}{A vector of weights, having the same length as the list.}
}
\value{
The weighted average of the series from the list.
}
\description{
Multiplies a list of multivariate time series with a weight vector and
sums up the result to get a weighted average.
}
