% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clusterEntropy_fast}
\alias{clusterEntropy_fast}
\title{Entropy for Clusterings}
\usage{
clusterEntropy_fast(assignments, normalize = FALSE)
}
\arguments{
\item{assignments}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters (code depends on this!).}

\item{normalize}{Should the entropy be normalized to [0,1]?}
}
\value{
The entropy as double.
}
\description{
Calculates the Shannon entropy for a cluster assignment vector. A value of
0 means that all elements are in one cluster, higher values indicate a more
even distribution of objects in the clusters. The value can be normalized
to [0,1] such that 1 means the same number of objects in each cluster.
}
