% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{compDist}
\alias{compDist}
\title{Compression-/Complexity-based Dissimilarity}
\usage{
compDist(x, y, symbolCount = 8, symbolLimits = NULL)
}
\arguments{
\item{x}{1st numeric vector/matrix (uni- or multi-variate time series).}

\item{y}{2nd numeric vector/matrix (uni- or multi-variate time series).}

\item{symbolCount}{Number of SAX symbols. Boundaries for the intervals will
be determined based on the standard normal distribution. As an alternative,
you can supply the boundaries directly.}

\item{symbolLimits}{Interval boundaries which will be used to convert the
time series to a SAX representation. Should be a monotonically increasing
vector starting with -Inf and ending with +Inf. The parameter
\code{symbolCount} is ignored if you supply a value here.}
}
\value{
The dissimilarity as numeric from the range [0,1].
}
\description{
Dissimilarity based on the length of the compressed single as well as concatenated
time series as described by Li et al. (2001). Time series are represented with
SAX first and then zipped, both according to Keogh et al. (2007). As an
improvement, the dissimilarity is scaled to the interval [0,1] (before: (0.5,1])
and made symmetric. Multi-variate time series are handled by attribute
concatenation.
}
\section{References}{


Keogh, E., Lonardi, S., Ratanamahatana, C. A., Wei, L., Lee, S.-H. & Handley, J.
(2007). Compression-based data mining of sequential data. \emph{Data Mining
and Knowledge Discovery, 14}(1), 99–129.

Li, M., Badger, J. H., Chen, X., Kwong, S., Kearney, P. & Zhang, H. (2001).
An information-based sequence distance and its application to whole
mitochondrial genome phylogeny. \emph{Bioinformatics, 17}(2), 149–154.
}

\seealso{
Other compression-based distances: \code{\link{compDistTSList}}
}
