% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{compDistTSList}
\alias{compDistTSList}
\title{Compression-/Complexity-based Dissimilarity}
\usage{
compDistTSList(tsList, symbolCount = 8, symbolLimits = NULL)
}
\arguments{
\item{tsList}{1) A list of numeric vectors/matrixes (uni- or multi-variate
time series). The dissimilarity of the list to itself (each time series to
each time series) will be computed, resulting in a symmetric dissimilarity
matrix. 2) A list with two components, each being a list of numeric vectors/
matrixes (uni- or multi-variate time series). The dissimilarity of each time
series from the 1st component to each time series from the 2nd component will
be computed.}

\item{symbolCount}{Number of SAX symbols. Boundaries for the intervals will
be determined based on the standard normal distribution. As an alternative,
you can supply the boundaries directly.}

\item{symbolLimits}{Interval boundaries which will be used to convert the
time series to a SAX representation. Should be a monotonically increasing
vector starting with -Inf and ending with +Inf. The parameter
\code{symbolCount} is ignored if you supply a value here.}
}
\value{
The dissimilarity matrix with each entry being a numeric from the
range [0,1].
}
\description{
Version of \code{\link{compDist}} which operates on a list of time series and
returns a distance matrix instead of a single distance. Saves computation time
compared to naive n^2 calls of the original function by computing the SAX
representations and single time series compression lengths only once for each
time series (not in every distance computation).
}
\section{References}{


Keogh, E., Lonardi, S., Ratanamahatana, C. A., Wei, L., Lee, S.-H. & Handley, J.
(2007). Compression-based data mining of sequential data. \emph{Data Mining
and Knowledge Discovery, 14}(1), 99–129.

Li, M., Badger, J. H., Chen, X., Kwong, S., Kearney, P. & Zhang, H. (2001).
An information-based sequence distance and its application to whole
mitochondrial genome phylogeny. \emph{Bioinformatics, 17}(2), 149–154.
}

\seealso{
Other compression-based distances: \code{\link{compDist}}
}
