% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{compDistTwoTSLists}
\alias{compDistTwoTSLists}
\title{Compression-/Complexity-based Dissimilarity for Two Lists}
\usage{
compDistTwoTSLists(tsList1, tsList2, symbolLimits)
}
\arguments{
\item{tsList1}{A list of numeric vectors/matrixes (uni- or multi-variate
time series).}

\item{tsList2}{A list of numeric vectors/matrixes (uni- or multi-variate
time series). Should have the same data structure as the first parameter.}

\item{symbolLimits}{Interval boundaries which will be used to convert the
time series to a SAX representation. Should be a monotonically increasing
vector starting with -Inf and ending with +Inf.}
}
\value{
The dissimilarity matrix with each entry being a numeric from the
range [0,1].
}
\description{
Version of \code{\link{compDistTSList}} which computes the compression-/
complexity-based dissimilarity from each time series of the first list to
each time series of the second list.
}
