% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conditionalEntropy_fast}
\alias{conditionalEntropy_fast}
\title{Conditional Entropy to Compare Clusterings}
\usage{
conditionalEntropy_fast(assignments, groundTruth, normalizeAndInvert = FALSE)
}
\arguments{
\item{assignments}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters (code depends on this!).}

\item{groundTruth}{Integer vector of true class labels containing only values
from 1 to k' with k' = number of classes.}

\item{normalizeAndInvert}{Should the entropy be normalized to [0,1] and inverted
such that high values indicate similar clusterings?}
}
\value{
The conditional entropy as double in [0, k'] (without normalization) or
a uniformity measure in [0,1] (with normalization).
}
\description{
Calculates the conditional Shannon entropy to compare two cluster assignment
vectors (external cluster validation). It is a value greater or equal 0,
lower values indicating more similarity (purer clusters).  Optionally, the
index can be normalized to [0,1] and we take 1 - normalized entropy to get
a uniformity measure where high values are good.
}
\details{
Be aware that this measure is asymmetric (classes are conditioned on/ analyzed
in) clusters and can still be high if the classes of the ground truth are
split up into multiple (but pure) clusters. Wu, Xiong and Chen (2009) propose
to use the symmetric variation of information (\code{\link{VI_fast}}) instead,
which is also based on entropy.

We use the base 2 logarithm for calculating entropy.
}
\section{References}{


Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{fowlkesMallows_fast}},
  \code{\link{pairCVIParameters_fast}},
  \code{\link{phi_fast}}, \code{\link{purity_fast}},
  \code{\link{randIndex_fast}},
  \code{\link{vanDongen_fast}}
}
