% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corDist_fast}
\alias{corDist_fast}
\title{(Fast) Correlation-based Dissimilarity}
\usage{
corDist_fast(x, y, beta = 0)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{beta}{If this parameter is smaller/equal zero, the formula \eqn{dist(x,y)=
\sqrt{2*(1-cor(x,y))}}{dist(x,y) = sqrt(2*(1-cor(x,y)))} is used (d1 in the
paper, equals \eqn{\frac{l2Dist(x,y)}{\sqrt{n}}}{l2Dist(x,y)/sqrt(n)} if time
series are z-standardized), otherwise
\eqn{dist(x,y)=\sqrt{(\frac{1-cor(x,y)}{1+cor(x,y)})^{\beta}}}{dist(x,y) =
sqrt(((1-cor(x,y))/(1+cor(x,y)))^beta)} (called d2 in the paper).}
}
\value{
The dissimilarity as double in the range [0,sqrt(2)] if \code{beta == 0}
and [0,Inf] otherwise. Is NaN if at least one series is constant.
}
\description{
Computes correlation-based dissimilarity as described by Golay et al. (1998).
The coding is inspired by the \code{TSclust::diss.cor()} method, but faster
because of the the C++ implementation.
}
\section{References}{


Golay, X., Kollias, S., Stoll, G., Meier, D., Valavanis, A. & Boesiger, P.
(1998). A new correlation-based fuzzy logic clustering algorithm for fmri.
\emph{Magnetic Resonance in Medicine, 40}(2), 249-260.
}

