% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cortFactor_fast}
\alias{cortFactor_fast}
\title{Temporal Correlation-Based Correction Factor for a Time Series Distance}
\usage{
cortFactor_fast(x, y, k = 2)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{k}{A non-negative constant for scaling.}
}
\value{
The correlation-based scaling factor as double from the range (0,2).
}
\description{
Considers the dissimilarity of two time series regarding their behavior,
namely if they move in the same direction (diff vectors used) at the
different points in time. This correlation between [-1,1] is scaled with
the exponential function to (0,2) (depending on \code{k}) and should be
multiplied with another dissimilarity to enhance it with this behavioral
information. Introduced by  Chouakria and Nagabhushan (2007).
}
\details{
This factor is currently integrated as a parameter into the L2 distance and
dynamic time warping distance of this package.
}
\section{References}{


Chouakria, A. D. & Nagabhushan, P. N. (2007). Adaptive dissimilarity index
for measuring time series proximity. \emph{Advances in Data Analysis and
Classification, 1}(1), 5-21.
}

