% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cortFactorMult_fast}
\alias{cortFactorMult_fast}
\title{Multi-variate Temporal Correlation-Based Correction Factor}
\usage{
cortFactorMult_fast(x, y, k = 2)
}
\arguments{
\item{x}{1st numeric matrix/multi-variate time series.}

\item{y}{2nd numeric matrix/multi-variate time series.}

\item{k}{A non-negative constant for scaling.}
}
\value{
The correlation-based scaling factor as double from the range (0,2).
}
\description{
Considers the dissimilarity of two time series regarding their behavior,
namely if they move in the same direction (diff vectors used) at the
different points in time. This correlation between [-1,1] is firstly
calculated for each dimension/attribute separately and then averaged.
It is scaled with the exponential function to (0,2) (depending on \code{k})
and should be multiplied with another dissimilarity to enhance it with
this behavioral information. Introduced by Chouakria and Nagabhushan (2007)
for the uni-variate case.
}
\details{
This factor is currently integrated as a parameter into the L2 distance and
dynamic time warping distance of this package.
}
\section{References}{


Chouakria, A. D. & Nagabhushan, P. N. (2007). Adaptive dissimilarity index
for measuring time series proximity. \emph{Advances in Data Analysis and
Classification, 1}(1), 5-21.

Kotsakos, D., Trajcevski, G., Gunopulos, D. & Aggarwal, C. C. (2014). Time-series
data clustering. In C. C. Aggarwal & C. K. Reddy (Eds.), \emph{Data clustering :
Algorithms and applications} (pp. 357–380). Chapman & Hall/CRC data mining and
knowledge discovery series. Boca Raton: CRC Press.
}

