% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{crossCorNormalized}
\alias{crossCorNormalized}
\title{Normalized Cross-Correlation Function}
\usage{
crossCorNormalized(x, y)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}
}
\value{
The normalized cross-correlation as vector (length equals
\emph{|x| + |y| - 1}.
}
\description{
Computes the cross-correlation function between two vectors and normalizes
it with the l2 norms of the vectors (so values are between -1 and 1). Internally,
convolution and FFT are used as described by Paparrizos and Gravano (2015).
}
\section{References}{

Paparrizos, J. & Gravano, L. (2015). K-shape: Efficient and accurate clustering
of time series. In \emph{Proceedings of the 2015 acm sigmod international
conference on management of data} (pp. 1855–1870). ACM.
}

\seealso{
Other cross-correlation functions: \code{\link{shapeBasedDistance}}
}
