% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DTWDist_fast}
\alias{DTWDist_fast}
\title{(Fast) Dynamic Time Warping Dissimilarity}
\usage{
DTWDist_fast(x, y, cid = FALSE, cortK = -1, normalize = FALSE)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{cid}{Should the distance be made "complexity-invariant"
(\code{\link{l2CompCorFactor_fast}})?}

\item{cortK}{Should the temporal behavior (correlation) of the time series'
diff vectors be considered (\code{\link{cortFactor_fast}})? A factor smaller
than 0 means no, higher factors will be used as parameter \code{k} in the
temporal correlation scaling function.}

\item{normalize}{Divide by the length of the longer time series (= minimum
amount of assignment steps) to account for series of different lengths in
your dataset.}
}
\value{
The distance as double.
}
\description{
Fast version of univariate dynamic time warping (unconstrained, symmetric1
step pattern) which uses a cyclic access strategy with a smaller cost matrix;
inspired by the C implementation of dynamic time warping in
\code{dtwclust::dtw_basic()}, but cuts even more overhead.
}
\details{
Be aware that it is not really a distance in the strict sense, as DTW
violates the triangle inequality.
}
\section{References}{

Berndt, D. J. & Clifford, J. (1994). Using dynamic time warping to find
patterns in time series. In \emph{Proceedings of the 3rd international
conference on knowledge discovery and data mining} (pp. 359–370). AAAI Press.
}

\seealso{
Other DTW functions: \code{\link{DTWDistMult_fast}},
  \code{\link{DTWDistSakoeChibaMult_fast}},
  \code{\link{DTWDistSakoeChiba_fast}}
}
