% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DTWDistSakoeChibaMult_fast}
\alias{DTWDistSakoeChibaMult_fast}
\title{(Fast) Multi-variate Dynamic Time Warping Dissimilarity with a Sakoe-Chiba Window}
\usage{
DTWDistSakoeChibaMult_fast(x, y, windowSize)
}
\arguments{
\item{x}{1st numeric matrix/multi-variate time series.}

\item{y}{2nd numeric matrix/multi-variate time series.}

\item{windowSize}{The maximum index difference which is considered when
matching elements. If greater/equal one, interpreted as absolute value.
If smaller then one, interpreted as fraction of the length of the longer
time series.}
}
\value{
The distance as double (not-a-number if matching is not possible
as the time series lengths differ more than \code{windowSize}).
}
\description{
Fast version of multivariate dynamic time warping (Sakoe-Chiba window as
constraint, symmetric1 step pattern, L2 distance for point-to-point comparisons)
which uses a cyclic access strategy with a smaller cost matrix; inspired by
the C implementation of dynamic time warping in \code{dtwclust::dtw_basic()},
but cuts even more overhead.
}
\details{
Be aware that it is not really a distance in the strict sense, as DTW
violates the triangle inequality.
}
\section{References}{

Berndt, D. J. & Clifford, J. (1994). Using dynamic time warping to find
patterns in time series. In \emph{Proceedings of the 3rd international
conference on knowledge discovery and data mining} (pp. 359–370). AAAI Press.

Sakoe, H., & Chiba, S. (1978). Dynamic programming algorithm optimization
for spoken word recognition. \emph{IEEE transactions on acoustics, speech,
and signal processing, 26}(1), 43-49.
}

\seealso{
Other DTW functions: \code{\link{DTWDistMult_fast}},
  \code{\link{DTWDistSakoeChiba_fast}},
  \code{\link{DTWDist_fast}}
}
