% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EDRDist_fast}
\alias{EDRDist_fast}
\title{(Fast) Edit Distance on Real Sequence}
\usage{
EDRDist_fast(x, y, epsilon, normalize = FALSE)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{epsilon}{Maximum distance between two time series elements to count a
match.}

\item{normalize}{Normalize the result to [0,1] considering the maximum
possible dissimilarity.}
}
\value{
The distance as double.
}
\description{
Computes the Edit distance on Real Sequence as described by Chen, Özsu
and Oria (2005). A match between two (real-valued) time series elements exists
if their L1 distance is below an \code{epsilon}. Apart from that, the
computation is similar to the standard edit distance. The coding is inspired
by the \code{TSdist::EDRDistance()} method, but faster because point-to-point
distances computation is integrated into the C++ code.
}
\details{
Despite the name, it is not really a distance in the strict sense, as EDR
violates the triangular inequality.
}
\section{References}{

Chen, L., Özsu, M. T. & Oria, V. (2005). Robust and fast similarity search
for moving object trajectories. In \emph{Proceedings of the 2005 acm sigmod
international conference on management of data} (pp. 491–502). ACM.
}

\seealso{
Other Edit distance functions: \code{\link{EDRDistMult_fast}},
  \code{\link{EDRDistSakoeChibaMult_fast}},
  \code{\link{EDRDistSakoeChiba_fast}},
  \code{\link{ERPDistMult_fast}},
  \code{\link{ERPDistSakoeChibaMult_fast}},
  \code{\link{ERPDistSakoeChiba_fast}},
  \code{\link{ERPDistSakoeChiba}},
  \code{\link{ERPDist_fast}}, \code{\link{ERPDist}}
}
