% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EDRDistSakoeChibaMult_fast}
\alias{EDRDistSakoeChibaMult_fast}
\title{(Fast) Multi-variate Edit Distance on Real Sequence and Sakoe-Chiba Window}
\usage{
EDRDistSakoeChibaMult_fast(x, y, epsilon, windowSize)
}
\arguments{
\item{x}{1st numeric matrix/multi-variate time series.}

\item{y}{2nd numeric matrix/multi-variate time series.}

\item{epsilon}{Maximum distance between two time series elements to count a
match.}

\item{windowSize}{The maximum index difference which is considered when
matching elements}
}
\value{
The distance as double (not-a-number if matching is not possible
as the time series lengths differ more than \code{windowSize}).
}
\description{
Computes the Edit distance on Real Sequence as described by Chen, Özsu
and Oria (2005), constraining the possible matches to a maximum index
difference of \code{windowSize} as desribed by Sakoe and Chiba (1978).
The coding is inspired by the \code{TSdist::EDRDistance()} method, but
faster because point-to-point distances computation is integrated into the
C++ code.
}
\details{
Despite the name, it is not really a distance in the strict sense, as EDR
violates the triangular inequality.
}
\section{References}{

Chen, L., Özsu, M. T. & Oria, V. (2005). Robust and fast similarity search
for moving object trajectories. In \emph{Proceedings of the 2005 acm sigmod
international conference on management of data} (pp. 491–502). ACM.

Sakoe, H., & Chiba, S. (1978). Dynamic programming algorithm optimization
for spoken word recognition. \emph{IEEE transactions on acoustics, speech,
and signal processing, 26}(1), 43-49.
}

\seealso{
Other Edit distance functions: \code{\link{EDRDistMult_fast}},
  \code{\link{EDRDistSakoeChiba_fast}},
  \code{\link{EDRDist_fast}},
  \code{\link{ERPDistMult_fast}},
  \code{\link{ERPDistSakoeChibaMult_fast}},
  \code{\link{ERPDistSakoeChiba_fast}},
  \code{\link{ERPDistSakoeChiba}},
  \code{\link{ERPDist_fast}}, \code{\link{ERPDist}}
}
