% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ERPDist}
\alias{ERPDist}
\title{(Fast) Edit Distance with Real Penalty}
\usage{
ERPDist(x, y, gapValue)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{gapValue}{If an element of one series is not matched to the other
series, its distance to the gapValue is computed instead (0 might be a
sensible default for standardized series).}
}
\value{
The distance as double.
}
\description{
Computes the Edit distance with real penalty as described by Chen and Ng
(2004). The coding is inspired by the \code{TSdist::ERPDistance()} method,
but faster because point-to-point distances computation is integrated into
the C++ code.
}
\section{References}{

Chen, L., & Ng, R. (2004, August). On the marriage of lp-norms and edit
distance. In \emph{Proceedings of the Thirtieth international conference
on Very large data bases-Volume 30} (pp. 792-803). VLDB Endowment.
}

\seealso{
Other Edit distance functions: \code{\link{EDRDistMult_fast}},
  \code{\link{EDRDistSakoeChibaMult_fast}},
  \code{\link{EDRDistSakoeChiba_fast}},
  \code{\link{EDRDist_fast}},
  \code{\link{ERPDistMult_fast}},
  \code{\link{ERPDistSakoeChibaMult_fast}},
  \code{\link{ERPDistSakoeChiba_fast}},
  \code{\link{ERPDistSakoeChiba}},
  \code{\link{ERPDist_fast}}
}
