% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ERPDist_fast}
\alias{ERPDist_fast}
\title{(Even Faster) Edit Distance with Real Penalty}
\usage{
ERPDist_fast(x, y, gapValue, normalize = FALSE)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{gapValue}{If an element of one series is not matched to the other
series, its distance to the gapValue is computed instead (0 might be a
sensible default for standardized series).}

\item{normalize}{Divide by the length of the longer time series (= minimum
amount of assignment steps) to account for series of different lengths in
your dataset.}
}
\value{
The distance as double.
}
\description{
Faster version of\link{ERPDist} which uses a cyclic access strategy
with a smaller cost matrix; inspired by the C implementation of dynamic
time warping in \code{dtwclust::dtw_basic()}.
}
\section{References}{

Chen, L., & Ng, R. (2004, August). On the marriage of lp-norms and edit
distance. In \emph{Proceedings of the Thirtieth international conference
on Very large data bases-Volume 30} (pp. 792-803). VLDB Endowment.
}

\seealso{
Other Edit distance functions: \code{\link{EDRDistMult_fast}},
  \code{\link{EDRDistSakoeChibaMult_fast}},
  \code{\link{EDRDistSakoeChiba_fast}},
  \code{\link{EDRDist_fast}},
  \code{\link{ERPDistMult_fast}},
  \code{\link{ERPDistSakoeChibaMult_fast}},
  \code{\link{ERPDistSakoeChiba_fast}},
  \code{\link{ERPDistSakoeChiba}}, \code{\link{ERPDist}}
}
