% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ERPDistSakoeChiba}
\alias{ERPDistSakoeChiba}
\title{(Fast) Edit Distance with Real Penalty and Sakoe-Chiba Window}
\usage{
ERPDistSakoeChiba(x, y, gapValue, windowSize)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{gapValue}{If an element of one series is not matched to the other
series, its distance to the gapValue is computed instead (0 might be a
sensible default for standardized series).}

\item{windowSize}{The maximum index difference which is considered when
matching elements.}
}
\value{
The distance as double (not-a-number if matching is not possible
as the time series lengths differ more than \code{windowSize}).
}
\description{
Computes the Edit distance with real penalty as described by Chen and Ng
(2004), constraining the possible matches to a maximum index difference of
\code{windowSize} as desribed by Sakoe and Chiba (1978). The coding is
inspired by the \code{TSdist::ERPDistance()} method, but faster because
point-to-point distances computation is integrated into the C++ code.
}
\section{References}{

Chen, L., & Ng, R. (2004, August). On the marriage of lp-norms and edit
distance. In \emph{Proceedings of the Thirtieth international conference
on Very large data bases-Volume 30} (pp. 792-803). VLDB Endowment.

Sakoe, H., & Chiba, S. (1978). Dynamic programming algorithm optimization
for spoken word recognition. \emph{IEEE transactions on acoustics, speech,
and signal processing, 26}(1), 43-49.
}

\seealso{
Other Edit distance functions: \code{\link{EDRDistMult_fast}},
  \code{\link{EDRDistSakoeChibaMult_fast}},
  \code{\link{EDRDistSakoeChiba_fast}},
  \code{\link{EDRDist_fast}},
  \code{\link{ERPDistMult_fast}},
  \code{\link{ERPDistSakoeChibaMult_fast}},
  \code{\link{ERPDistSakoeChiba_fast}},
  \code{\link{ERPDist_fast}}, \code{\link{ERPDist}}
}
