% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FastTSDistances}
\alias{FastTSDistances}
\alias{FastTSDistances-package}
\title{Fast Distance Measures for Time Series}
\description{
Fast implementations of distance measures for time series.
}
\author{
Jakob Bach <jakob.bach@student.kit.edu>
}
