% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fowlkesMallows_fast}
\alias{fowlkesMallows_fast}
\title{Fowlkes-Mallows Index}
\usage{
fowlkesMallows_fast(pairCVIParams, normalize = FALSE)
}
\arguments{
\item{pairCVIParams}{Output of \code{link{pairCVIParameters}} which has to be
called with the two cluster assignment vector to be compared.}

\item{normalize}{Should the Fowlkes-Mallows Index be corrected for chance?}
}
\value{
The Fowlkes-Mallows Index as double (at most 1 for identical clusterings,
normal Fowlkes-Mallows greater than zero, normalized one can also be negative).
}
\description{
Calculates the index of Fowlkes and Mallows (1983) to compare two cluster
assignment vectors (external cluster validation). It is a value in [0,1],
the geometric mean of precision and recall, higher values indicating more
similarity. The index can be corrected for similarity by chance as proposed
by Wu, Xiong and Chen (2009), then also possibly yielding negative results
(if the similarity is worse than random assignment) while the maximum is still
1 (and values are usually positive).
}
\section{References}{


Fowlkes, E. B. & Mallows, C. L. (1983). A method for comparing two hierarchical
clusterings. \emph{Journal of the American statistical association, 78}(383),
553-569.

Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{conditionalEntropy_fast}},
  \code{\link{pairCVIParameters_fast}},
  \code{\link{phi_fast}}, \code{\link{purity_fast}},
  \code{\link{randIndex_fast}},
  \code{\link{vanDongen_fast}}
}
