% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generalizedDB_fast}
\alias{generalizedDB_fast}
\title{Generalized Davies-Bouldin Index}
\usage{
generalizedDB_fast(interClusterDistances, intraClusterDistances)
}
\arguments{
\item{interClusterDistances}{A matrix representing the distances between
clusters (separation), with the number of rows/columns equal to the number
of clusters.}

\item{intraClusterDistances}{A vector representing the distances within a
cluster (compactness), with its length equal to the number of clusters.}
}
\value{
The Generalized Davies-Bouldin Index. Could be Inf or NaN if there
are clusters with a distance of zero between them (which is a bad clustering
result).
}
\description{
Calculates a generalized version of the Davies-Bouldin Index for internal
cluster validation. The index is expressed by a ratio of cluster compactness
and cluster separation, summed and averaged over all clusters. This generalized
method does not define the concrete way to compute distances within clusters
and between clusters, but simply takes these distances as input to compute the
Davies-Bouldin Index. Lower values indicate better clustering quality.
}
\section{References}{


Davies, D. L. & Bouldin, D. W. (1979). A cluster separation measure. \emph{IEEE
transactions on pattern analysis and machine intelligence, 1}(2), 224–227.
}

\seealso{
Other Internal Cluster Validity Indices: \code{\link{generalizedDunn_fast}},
  \code{\link{iGeneralizedDB_fast}}
}
