% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generalizedDunn_fast}
\alias{generalizedDunn_fast}
\title{Generalized Dunn Index}
\usage{
generalizedDunn_fast(interClusterDistances, intraClusterDistances)
}
\arguments{
\item{interClusterDistances}{A symmetric matrix representing the distances
between clusters (separation), with the number of rows/columns equal to the
number of clusters.}

\item{intraClusterDistances}{A vector representing the distances within a
cluster (compactness), with its length equal to the number of clusters.}
}
\value{
The Generalized Dunn Index. Could be Inf or NaN if there are clusters
with a distance of zero between them (which is a bad clustering result).
}
\description{
Calculates a generalized version of the Dunn Index, allowing an arbitrary
measure of cluster separation (which goes to the numerator and is minimized)
and an arbitrary measure of cluster compactness (which goes to the denominator
and is maximized over all clusters). Dunn used the highly outlier-prone
single linkage measure for separation and complete linkage for compactness.
Higher values indicate better clustering quality.
}
\section{References}{


Bezdek, J. C. & Pal, N. R. (1998). Some new indexes of cluster validity.
\emph{IEEE Transactions on Systems, Man, and Cybernetics, Part B (Cybernetics),
28}(3), 301–315.

Dunn, J. C. (1973). A fuzzy relative of the isodata process and its use in
detecting compact well-separated clusters. \emph{Journal of Cybernetics, 3}(3),
32–57.
}

\seealso{
Other Internal Cluster Validity Indices: \code{\link{generalizedDB_fast}},
  \code{\link{iGeneralizedDB_fast}}
}
