% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iGeneralizedDB_fast}
\alias{iGeneralizedDB_fast}
\title{Inverted Generalized Davies-Bouldin Index}
\usage{
iGeneralizedDB_fast(interClusterDistances, intraClusterDistances)
}
\arguments{
\item{interClusterDistances}{A matrix representing the distances between
clusters (separation), with the number of rows/columns equal to the number
of clusters.}

\item{intraClusterDistances}{A vector representing the distances within a
cluster (compactness), with its length equal to the number of clusters.}
}
\value{
The Inverted Generalized Davies-Bouldin Index. Could be Inf or NaN
if there are clusters with a distance of zero between them (which is a bad
clustering result) or infinity between them (which might indicate an error
in the computation of your dissimilarity matrix).
}
\description{
Calculates a generalized version of the Davies-Bouldin Index, similar to
\code{\link{generalizedDB_fast}}. The only difference is that the separation
measure is divided by the compactness measure (inverted compared to original
index), so high values are desirable.
}
\section{References}{


Davies, D. L. & Bouldin, D. W. (1979). A cluster separation measure. \emph{IEEE
transactions on pattern analysis and machine intelligence, 1}(2), 224–227.
}

\seealso{
Other Internal Cluster Validity Indices: \code{\link{generalizedDB_fast}},
  \code{\link{generalizedDunn_fast}}
}
