% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l1Dist_fast}
\alias{l1Dist_fast}
\title{(Fast) L1 Distance}
\usage{
l1Dist_fast(x, y)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}
}
\value{
The distance as double.
}
\description{
Computes the standard Manhattan distance with a fast C++ implementation.
}
\seealso{
Other L_p distances: \code{\link{l1DistMult_fast}},
  \code{\link{l2DistMult_fast}}, \code{\link{l2Dist_fast}},
  \code{\link{l2Dist}}, \code{\link{l2Norm_fast}},
  \code{\link{lmaxDistMult_fast}},
  \code{\link{lmaxDist_fast}}
}
