% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2CompCorFactorMult_fast}
\alias{l2CompCorFactorMult_fast}
\title{Multi-variate L2 Complexity Correction Factor}
\usage{
l2CompCorFactorMult_fast(x, y)
}
\arguments{
\item{x}{1st numeric matrix/multi-variate time series.}

\item{y}{2nd numeric vector/multi-variate time series.}
}
\value{
The complexity correction factor as double. Is infinity if one series
is constant in all attributes and the other one not.
}
\description{
Calculates the complexity correction factor for the distance between two
multi-variate time series, using the L2 norm of each time series' attributes'
diff vectors as complexity estimate. Can be combined with any distance as a
scaling factor (distances between vectors of different complexity become more
prominent). Does not obey the triangular equality if its combined with the
Euclidean distance, but a relaxed version (see reference).
}
\details{
This factor is currently integrated as a parameter into the L2 distance and
dynamic time warping distance of this package.
}
\section{References}{


Batista, G. E., Keogh, E. J., Tataw, O. M. & De Souza, V. M. (2014). Cid: An
efficient complexity-invariant distance for time series. \emph{Data Mining and
Knowledge Discovery, 28}(3), 634-669.

Kotsakos, D., Trajcevski, G., Gunopulos, D. & Aggarwal, C. C. (2014). Time-series
data clustering. In C. C. Aggarwal & C. K. Reddy (Eds.), \emph{Data clustering :
Algorithms and applications} (pp. 357–380). Chapman & Hall/CRC data mining and
knowledge discovery series. Boca Raton: CRC Press.
}

\seealso{
\code{\link{l2Dist_fast}}, \code{\link{DTWDist_fast}}
}
