% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{l2Dist}
\alias{l2Dist}
\title{L2 Distance}
\usage{
l2Dist(x, y)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}
}
\value{
The L2 norm as numeric.
}
\description{
Computes the standard Euclidean distance.
}
\seealso{
Other L_p distances: \code{\link{l1DistMult_fast}},
  \code{\link{l1Dist_fast}}, \code{\link{l2DistMult_fast}},
  \code{\link{l2Dist_fast}}, \code{\link{l2Norm_fast}},
  \code{\link{lmaxDistMult_fast}},
  \code{\link{lmaxDist_fast}}
}
