% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2Dist_fast}
\alias{l2Dist_fast}
\title{(Fast) L2 Distance}
\usage{
l2Dist_fast(x, y, cid = FALSE, cortK = -1)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}

\item{cid}{Should the distance be made complexity invariant
(\code{\link{l2CompCorFactor_fast}})?}

\item{cortK}{Should the temporal behavior (correlation) of the time series'
diff vectors be considered (\code{\link{cortFactor_fast}})? A factor smaller
than 0 means no, higher factors will be used as parameter \code{k} in the
temporal correlation scaling function.}
}
\value{
The distance as double.
}
\description{
Computes the standard Euclidean distance with a fast C++ implementation.
}
\seealso{
Other L_p distances: \code{\link{l1DistMult_fast}},
  \code{\link{l1Dist_fast}}, \code{\link{l2DistMult_fast}},
  \code{\link{l2Dist}}, \code{\link{l2Norm_fast}},
  \code{\link{lmaxDistMult_fast}},
  \code{\link{lmaxDist_fast}}
}
