% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2Norm_fast}
\alias{l2Norm_fast}
\title{(Fast) L2 Norm}
\usage{
l2Norm_fast(x)
}
\arguments{
\item{x}{A numeric vector/time series.}
}
\value{
The norm as double.
}
\description{
Computes the standard Euclidean norm with a fast C++ implementation.
}
\seealso{
Other L_p distances: \code{\link{l1DistMult_fast}},
  \code{\link{l1Dist_fast}}, \code{\link{l2DistMult_fast}},
  \code{\link{l2Dist_fast}}, \code{\link{l2Dist}},
  \code{\link{lmaxDistMult_fast}},
  \code{\link{lmaxDist_fast}}
}
