% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmaxDist_fast}
\alias{lmaxDist_fast}
\title{(Fast) Chebyshev Distance}
\usage{
lmaxDist_fast(x, y)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}
}
\value{
The distance as double.
}
\description{
Computes the standard Chebyshev distance (maximum metric) with a fast C++
implementation.
}
\seealso{
Other L_p distances: \code{\link{l1DistMult_fast}},
  \code{\link{l1Dist_fast}}, \code{\link{l2DistMult_fast}},
  \code{\link{l2Dist_fast}}, \code{\link{l2Dist}},
  \code{\link{l2Norm_fast}},
  \code{\link{lmaxDistMult_fast}}
}
