% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{minMaxNormalize}
\alias{minMaxNormalize}
\title{Min-Max Normalization}
\usage{
minMaxNormalize(x)
}
\arguments{
\item{x}{A numeric vector/matrix (uni- or multivariate time series).}
}
\value{
Normalized time series.
}
\description{
Normalizes a time series to [0,1]. Multi-variate time series (matrices) are
normalized per column.
}
