% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PAA_fast}
\alias{PAA_fast}
\title{Piecewise Aggregate Approximation}
\usage{
PAA_fast(x, windowCount)
}
\arguments{
\item{x}{Numeric vector/time series.}

\item{windowCount}{The number of windows for the shortened time series.}
}
\value{
A numeric vector of the length \code{windowCount}.
}
\description{
Divides a time series into \code{windowCount} frames of equal length and
represents each interval by its mean. If the time series length is not
divisible by the number of windows, one element might belong to two windows,
but only with a certain fraction to each. Inspired by the pure R implementation
in \code{TSclust::PAA()}, but faster because C++ code.
}
\section{References}{


Keogh, E. J. & Pazzani, M. J. (2000). Scaling up dynamic time warping for
datamining applications. In \emph{Proceedings of the sixth acm sigkdd
international conference on knowledge discovery and data mining} (pp. 285-289).
ACM.

Keogh, E., Chakrabarti, K., Pazzani, M. & Mehrotra, S. (2001). Dimensionality
reduction for fast similarity search in large time series databases.
\emph{Knowledge and information Systems}, 3(3), 263-286.
}

\seealso{
Other piecewise aggregation functions: \code{\link{PKurtAA_fast}},
  \code{\link{PMaxAA_fast}}, \code{\link{PMedAA_fast}},
  \code{\link{PMinAA_fast}}, \code{\link{PSDAA_fast}},
  \code{\link{PSkewAA_fast}}
}
