% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairCVIParameters_fast}
\alias{pairCVIParameters_fast}
\title{Statistics for External CVIs based on Pairwise Comparison}
\usage{
pairCVIParameters_fast(assignments1, assignments2)
}
\arguments{
\item{assignments1}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters (code depends on this!).}

\item{assignments2}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters.}
}
\value{
Vector with the four components called \code{m, m1, m2, M} by Wu,
Xiong and Chen (2009).
}
\description{
Calculates the summary statistics [m, m1, m2, M] which can be used to compute
multiple normalized external CVIs based on the formulas of Wu, Xiong and Chen (2009).
}
\section{References}{


Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{conditionalEntropy_fast}},
  \code{\link{fowlkesMallows_fast}},
  \code{\link{phi_fast}}, \code{\link{purity_fast}},
  \code{\link{randIndex_fast}},
  \code{\link{vanDongen_fast}}
}
