% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcDistOneTSList}
\alias{pdcDistOneTSList}
\title{Permutation Distribution Distance for One List of Time Series}
\usage{
pdcDistOneTSList(tsList, subSequenceLength = NULL)
}
\arguments{
\item{tsList}{A list of numeric vectors (uni-variate time series). Each time
series should have at least two elements.}

\item{subSequenceLength}{Number of elements which form each subsequence. Will
be determined as integer in [2,7] by a heuristic if not provided.}
}
\value{
The distance matrix with each entry being a numeric from the range [0,4].
}
\description{
Version of \code{\link{pdcDistTSList}} which computes the PDC distance from
each time series of a list to each time series of the same list.
}
