% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcDistTSList}
\alias{pdcDistTSList}
\title{Permutation Distribution Distance for a List of Time Series}
\usage{
pdcDistTSList(tsList, subSequenceLength = NULL)
}
\arguments{
\item{tsList}{1) A list of numeric vectors (uni-variate time series). The
dissimilarity of the list to itself (each time series to each time series)
will be computed, resulting in a symmetric dissimilarity matrix. 2) A list
with two components, each being a list of numeric vectors (uni-variate time
series). The dissimilarity of each time series from the 1st component to each
time series from the 2nd component will be computed. The entropy heuristic is
only computed with the second component (for comparability to previously
computed distances). 1+2) Each time series should have at least two elements.}

\item{subSequenceLength}{Number of elements which form each subsequence. Will
be determined as integer in [2,7] by a heuristic if not provided.}
}
\value{
The distance matrix with each entry being a numeric from the range [0,4].
}
\description{
Simplified version of the distance used for permutation distribution clustering
as described by Brandmaier (2015). Shortens the high-level code a bit and
handles some errors for short time series.
}
\section{References}{


Brandmaier, A. M. (2011). \emph{Permutation distribution clustering and structural
equation model trees} (Doctoral dissertation, Universität des Saarlandes, Saarbrücken).

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of
Time Series. \emph{Journal of Statistical Software, 67}(5), 1-23.
}

\seealso{
Other PDC functions: \code{\link{pdcDistTSListMult}},
  \code{\link{pdcDistTwoTSMult}},
  \code{\link{pdcDistTwoTS}},
  \code{\link{pdcEntropyHeuristicMult}},
  \code{\link{pdcEntropyHeuristic}},
  \code{\link{pdcEntropy}}
}
