% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcDistTwoTS}
\alias{pdcDistTwoTS}
\title{Permutation Distribution Distance for Two Time Series}
\usage{
pdcDistTwoTS(x, y, subSequenceLength = NULL)
}
\arguments{
\item{x}{1st numeric vector/time series (minimum length: 2 elements).}

\item{y}{2nd numeric vector/time series (minimum length: 2 elements).}

\item{subSequenceLength}{Number of elements which form each subsequence. Will
be determined as integer in [2,7] by a heuristic if not provided.}
}
\value{
The distance as double from the range [0,4].
}
\description{
Simplified version of the distance used for permutation distribution clustering
as described by Brandmaier (2015). Shortens the high-level code a bit for the two
time series case and handles some errors for short time series.
}
\section{References}{


Brandmaier, A. M. (2011). \emph{Permutation distribution clustering and structural
equation model trees} (Doctoral dissertation, Universität des Saarlandes, Saarbrücken).

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of
Time Series. \emph{Journal of Statistical Software, 67}(5), 1-23.
}

\seealso{
Other PDC functions: \code{\link{pdcDistTSListMult}},
  \code{\link{pdcDistTSList}},
  \code{\link{pdcDistTwoTSMult}},
  \code{\link{pdcEntropyHeuristicMult}},
  \code{\link{pdcEntropyHeuristic}},
  \code{\link{pdcEntropy}}
}
