% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcDistTwoTSLists}
\alias{pdcDistTwoTSLists}
\title{Permutation Distribution Distance for Two Lists of Time Series}
\usage{
pdcDistTwoTSLists(tsList1, tsList2, subSequenceLength = NULL)
}
\arguments{
\item{tsList1}{A list of numeric vectors (uni-variate time series). Each time
series should have at least two elements.}

\item{tsList2}{A list of numeric vectors (uni-variate time series). Each time
series should have at least two elements. Is used for the entropy heuristic
if no subSequenceLength is provided,}

\item{subSequenceLength}{Number of elements which form each subsequence. Will
be determined as integer in [2,7] by a heuristic if not provided.}
}
\value{
The distance matrix with each entry being a numeric from the range [0,4].
}
\description{
Version of \code{\link{pdcDistTSList}} which computes the PDC distance from
each time series of the first list to each time series of the second list.
}
