% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcDistTwoTSListsMult}
\alias{pdcDistTwoTSListsMult}
\title{Multi-variate Permutation Distribution Distance for Two Lists of Time Series}
\usage{
pdcDistTwoTSListsMult(tsList1, tsList2, subSequenceLength = NULL)
}
\arguments{
\item{tsList1}{A list of numeric matrixes (multi-variate time series). Each
time series should have at least two elements.}

\item{tsList2}{A list of numeric metrixes (multi-variate time series). Each
time series should have at least two elements and the same number of attributes
as the series from the first list. Is used for the entropy heuristic if no
subSequenceLength is provided.}

\item{subSequenceLength}{Number of elements which form each subsequence. Will
be determined as integer in [2,7] by a heuristic if not provided.}
}
\value{
The distance matrix with each entry being a numeric from the range
[0,4*attributeCount].
}
\description{
Version of \code{\link{pdcDistTSListMult}} which computes the PDC distance from
each time series of the first list to each time series of the second list.
}
