% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcDistTwoTSMult}
\alias{pdcDistTwoTSMult}
\title{Multi-variate Permutation Distribution Distance for Two Time Series}
\usage{
pdcDistTwoTSMult(x, y, subSequenceLength = NULL)
}
\arguments{
\item{x}{1st numeric matrix (multi-variate time series; minimum number of rows: 2).}

\item{y}{2nd numeric matrix (multi-variate time series; minimum number of rows: 2;
needs to have the same number of columns as the first time series).}

\item{subSequenceLength}{Number of elements which form each subsequence. Will
be determined as integer in [2,7] by a heuristic if not provided.}
}
\value{
The distance as double from the range [0,4*attributeCount].
}
\description{
Multi-variate version of \code{\link{pdcDistTwoTS}} which calculates the pdc
distance for each attribute and then takes the l2 norm of the resulting vector
as proposed by Brandmaier (2011, p. 18).
}
\section{References}{


Brandmaier, A. M. (2011). \emph{Permutation distribution clustering and structural
equation model trees} (Doctoral dissertation, Universität des Saarlandes, Saarbrücken).

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of
Time Series. \emph{Journal of Statistical Software, 67}(5), 1-23.
}

\seealso{
Other PDC functions: \code{\link{pdcDistTSListMult}},
  \code{\link{pdcDistTSList}}, \code{\link{pdcDistTwoTS}},
  \code{\link{pdcEntropyHeuristicMult}},
  \code{\link{pdcEntropyHeuristic}},
  \code{\link{pdcEntropy}}
}
