% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcEntropy}
\alias{pdcEntropy}
\title{Permutation Distribution Entropy}
\usage{
pdcEntropy(timeSeries, subSequenceLength)
}
\arguments{
\item{timeSeries}{A numeric vector/time series (minimum length: 2 elements).}

\item{subSequenceLength}{Number of elements in subsequences which are used to
count patterns.}
}
\value{
The normalized entropy as double.
}
\description{
More stable version of the entropy measure which is used by Brandmaier (2015)
to determine the optimal embedding length in permutation distribution clustering.
Mainly copies the (internal) function \code{pdc::codebook.entropy}, but also
handles the case where only one pattern occurs in the time series.
}
\details{
Encodes time series subsequences of the requested length, counts the relative
frequency of each ordinal pattern and computes the entropy on this distribution.
To obtain an unbiased estimate, Brandmaier proposes to divide by the log of
non-zero-bin count.
}
\section{References}{


Brandmaier, A. M. (2011). \emph{Permutation distribution clustering and structural
equation model trees} (Doctoral dissertation, Universität des Saarlandes, Saarbrücken).

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of
Time Series. \emph{Journal of Statistical Software, 67}(5), 1-23.
}

\seealso{
Other PDC functions: \code{\link{pdcDistTSListMult}},
  \code{\link{pdcDistTSList}},
  \code{\link{pdcDistTwoTSMult}},
  \code{\link{pdcDistTwoTS}},
  \code{\link{pdcEntropyHeuristicMult}},
  \code{\link{pdcEntropyHeuristic}}
}
