% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcEntropyHeuristic}
\alias{pdcEntropyHeuristic}
\title{Permutation Distribution Entropy Heuristic}
\usage{
pdcEntropyHeuristic(tsList)
}
\arguments{
\item{tsList}{A list of numeric vectors/time series (which should each have a
minimum length of two).}
}
\value{
The optimal subsequence/pattern length as integer from the range [2,7].
}
\description{
Simplified and stabilized version of the entropy heuristic used by Brandmaier (2015)
(\code{pdc::entropyHeuristic}) to determine the subsequence length in permutation
distribution clustering. Shortens the code a bit and can handle time series
with only one permutation pattern.
}
\section{References}{


Brandmaier, A. M. (2011). \emph{Permutation distribution clustering and structural
equation model trees} (Doctoral dissertation, Universität des Saarlandes, Saarbrücken).

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of
Time Series. \emph{Journal of Statistical Software, 67}(5), 1-23.
}

\seealso{
Other PDC functions: \code{\link{pdcDistTSListMult}},
  \code{\link{pdcDistTSList}},
  \code{\link{pdcDistTwoTSMult}},
  \code{\link{pdcDistTwoTS}},
  \code{\link{pdcEntropyHeuristicMult}},
  \code{\link{pdcEntropy}}
}
