% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{pdcEntropyHeuristicMult}
\alias{pdcEntropyHeuristicMult}
\title{Multi-variate Permutation Distribution Entropy Heuristic}
\usage{
pdcEntropyHeuristicMult(tsList)
}
\arguments{
\item{tsList}{A list of numeric vectors and/or matrixes which represent uni-variate
and/or multi-variate time series.}
}
\value{
The optimal subsequence/pattern length as integer from the range [2,7].
}
\description{
Multi-variate version of \code{\link{pdcEntropyHeuristic}} which splits
multi-variate time series (matrices) in multiple uni-variate time series
(vectors) before applying the heuristic. This approach was proposed in
the dissertation of Brandmaier (2011, p. 25).
}
\section{References}{


Brandmaier, A. M. (2011). \emph{Permutation distribution clustering and structural
equation model trees} (Doctoral dissertation, Universität des Saarlandes, Saarbrücken).

Brandmaier, A. M. (2015). pdc: An R Package for Complexity-Based Clustering of
Time Series. \emph{Journal of Statistical Software, 67}(5), 1-23.
}

\seealso{
Other PDC functions: \code{\link{pdcDistTSListMult}},
  \code{\link{pdcDistTSList}},
  \code{\link{pdcDistTwoTSMult}},
  \code{\link{pdcDistTwoTS}},
  \code{\link{pdcEntropyHeuristic}},
  \code{\link{pdcEntropy}}
}
