% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phi_fast}
\alias{phi_fast}
\title{Phi Coefficient}
\usage{
phi_fast(pairCVIParams)
}
\arguments{
\item{pairCVIParams}{Output of \code{link{pairCVIParameters}} which has to be
called with the two cluster assignment vector to be compared.}
}
\value{
Phi/Gamma coefficient as double from the range (-1,1].
}
\description{
Calculates the Phi coefficient of Pearson (1900) to compare two cluster
assignment vectors (external cluster validation). It is a correlation value,
therefore being in the interval (-1,1], higher values indicating more similar
clusterings. The name varies in the literature. Instead of Phi, some sources
also call it Gamma. In Wu, Xiong and Chen (2009) its called "Hubert's Gamma
statistic I".
}
\section{References}{


Pearson, K. (1900). Mathematical contributions to the theory of evolution. vii.
on the correlation of characters not quantitatively measurable. \emph{Philosophical
Transactions of the Royal Society of London. Series A, Containing Papers of a
Mathematical or Physical Character}, 195, 1-405.

Pearson, K. & Heron, D. (1913). On theories of association. \emph{Biometrika},
9(1/2), 159-315.

Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{conditionalEntropy_fast}},
  \code{\link{fowlkesMallows_fast}},
  \code{\link{pairCVIParameters_fast}},
  \code{\link{purity_fast}}, \code{\link{randIndex_fast}},
  \code{\link{vanDongen_fast}}
}
