% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PMaxAA_fast}
\alias{PMaxAA_fast}
\title{Piecewise Maximum Aggregate Approximation}
\usage{
PMaxAA_fast(x, windowCount)
}
\arguments{
\item{x}{Numeric vector/time series.}

\item{windowCount}{The number of windows for the shortened time series.}
}
\value{
A numeric vector of the length \code{windowCount}.
}
\description{
Divides a time series into \code{windowCount} frames of equal length and
represents each interval by its maximum. If the time series length is not
divisible by the number of windows, elements still assigned uniquely to one
window each.
}
\details{
The function is similar to \code{\link{PAA_fast}}, but simply uses a different
aggregate.
}
\seealso{
Other piecewise aggregation functions: \code{\link{PAA_fast}},
  \code{\link{PKurtAA_fast}}, \code{\link{PMedAA_fast}},
  \code{\link{PMinAA_fast}}, \code{\link{PSDAA_fast}},
  \code{\link{PSkewAA_fast}}
}
