% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PSDAA_fast}
\alias{PSDAA_fast}
\title{Piecewise Standard Deviation Aggregate Approximation}
\usage{
PSDAA_fast(x, windowCount, sample = FALSE)
}
\arguments{
\item{x}{Numeric vector/time series.}

\item{windowCount}{The number of windows for the shortened time series.}

\item{sample}{Compute sample standard deviation instead of population
standard deviation (divide by n-1 instead of n).}
}
\value{
A numeric vector of the length \code{windowCount}.
}
\description{
Divides a time series into \code{windowCount} frames of equal length and
represents each interval by its standard deviation. If the time series length
is not divisible by the number of windows, one element might belong to two
windows, but only with a certain fraction to each.
}
\details{
The function is similar to \code{\link{PAA_fast}}, but simply uses a different
aggregate.
}
\seealso{
Other piecewise aggregation functions: \code{\link{PAA_fast}},
  \code{\link{PKurtAA_fast}}, \code{\link{PMaxAA_fast}},
  \code{\link{PMedAA_fast}}, \code{\link{PMinAA_fast}},
  \code{\link{PSkewAA_fast}}
}
