% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{purity_fast}
\alias{purity_fast}
\title{Purity Measure}
\usage{
purity_fast(assignments, groundTruth)
}
\arguments{
\item{assignments}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters (code depends on this!).}

\item{groundTruth}{Integer vector of class (ground truth) assignments containing
only values from 1 to k with k = number of clusters.}
}
\value{
The purity measure as double in (0,1].
}
\description{
Calculates the purity measure (e.g. described by Wu, Xiong and Chen (2009))
to compare two cluster assignment vectors (external cluster validation). It
is a value in (0,1], higher values indicating more similarity. It finds the
most common ground truth class in each cluster and sums over these relative
frequencies.
}
\details{
Be aware that this measure is asymmetric and can still be high if the classes
of the ground truth are split up into multiple (but pure) clusters. Wu, Xiong
and Chen (2009) propose to use the symmetric van Dongen measure
(\code{\link{vanDongen_fast}}) instead.
}
\section{References}{


Van Dongen, S. (2000). \emph{Performance criteria for graph clustering and markov
cluster experiments}. National Research Institute for Mathematics and Computer
Science. Amsterdam.

Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{conditionalEntropy_fast}},
  \code{\link{fowlkesMallows_fast}},
  \code{\link{pairCVIParameters_fast}},
  \code{\link{phi_fast}}, \code{\link{randIndex_fast}},
  \code{\link{vanDongen_fast}}
}
