% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randIndex_fast}
\alias{randIndex_fast}
\title{Rand Index}
\usage{
randIndex_fast(pairCVIParams, normalize = FALSE)
}
\arguments{
\item{pairCVIParams}{Output of \code{link{pairCVIParameters}} which has to be
called with the two cluster assignment vector to be compared.}

\item{normalize}{Should the Rand Index be corrected for chance? (Adjusted Rand
Index as proposed by Hubert and Arabie (1985))}
}
\value{
The (Adjusted) Rand Index as double (at most 1 for identical clusterings,
normal Rand Index greater than zero, adjusted one can also be negative).
}
\description{
Calculates the Rand Index of Rand (1971) to compare two cluster assignment
vectors (external cluster validation). It is a value in (0,1], higher values
indicating more similarity. The index can be corrected for similarity by
chance as proposed by Hubert and Arabie (1985), then also possibly yielding
negative results (if the similarity is worse than random assignment) while
the maximum is still 1 (and values are usually positive).
}
\section{References}{


Hubert, L. & Arabie, P. (1985). Comparing partitions. \emph{Journal of
classification,2}(1), 193-218.

Rand, W. M. (1971). Objective criteria for the evaluation of clustering methods.
\emph{Journal of the American Statistical association, 66}(336), 846-850.

Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{conditionalEntropy_fast}},
  \code{\link{fowlkesMallows_fast}},
  \code{\link{pairCVIParameters_fast}},
  \code{\link{phi_fast}}, \code{\link{purity_fast}},
  \code{\link{vanDongen_fast}}
}
