% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{SAXLimitsDataAdaptive}
\alias{SAXLimitsDataAdaptive}
\title{Symbolic Aggregate Aproximation Limits (Flexible)}
\usage{
SAXLimitsDataAdaptive(tsList, alphabetSize = 8)
}
\arguments{
\item{tsList}{A list of numeric vectors/matrixes (uni- or multi-variate time
series).}

\item{alphabetSize}{Number of symbols (intervals).}
}
\value{
The interval boundaries as monotonically increasing vector from negative
infinity to positive infinity.
}
\description{
Calculates the interval boundaries for a SAX representation based on the real
distribution of values in a time series list. Uses evenly-spaced quantiles on
the real data instead of the normal distribution proposed by the original authors.
}
\section{References}{


Lin, J., Keogh, E., Lonardi, S. & Chiu, B. (2003). A symbolic representation
of time series, with implications for streaming algorithms. In \emph{Proceedings
of the 8th acm sigmod workshop on research issues in data mining and knowledge
discovery} (pp. 2-11). ACM.
}

\seealso{
Other SAX functions: \code{\link{SAXLimitsOriginal}},
  \code{\link{SAX_fast}}
}
