% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{SAXLimitsOriginal}
\alias{SAXLimitsOriginal}
\title{Symbolic Aggregate Aproximation Limits (Standard)}
\usage{
SAXLimitsOriginal(alphabetSize = 8)
}
\arguments{
\item{alphabetSize}{Number of symbols (intervals).}
}
\value{
The interval boundaries as monotonically increasing vector from negative
infinity to positive infinity.
}
\description{
Calculates the interval boundaries for a SAX representation based on the normal
distribution. Same as \code{TSclust::SAX.breakpoints.table()} (not exported
function). This is the approach of the original authors.
}
\section{References}{


Lin, J., Keogh, E., Lonardi, S. & Chiu, B. (2003). A symbolic representation
of time series, with implications for streaming algorithms. In \emph{Proceedings
of the 8th acm sigmod workshop on research issues in data mining and knowledge
discovery} (pp. 2-11). ACM.
}

\seealso{
Other SAX functions: \code{\link{SAXLimitsDataAdaptive}},
  \code{\link{SAX_fast}}
}
