% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{shapeBasedDistance}
\alias{shapeBasedDistance}
\title{Shape-Based Distance}
\usage{
shapeBasedDistance(x, y)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}
}
\value{
The dissimilarity as numeric from the range [0,2].
}
\description{
Computes the dissimilarity as \emph{1 - maximum normalized cross-correlation
coefficient} as described by Paparrizos and Gravano (2015). Multi-variate
time series are handled by calculating the cross-correlation between corresponding
attributes in \code{x} and \code{y}, averaging over attributes and then taking
the average cross-correlation at the lag which maximizes it.
}
\section{References}{

Paparrizos, J. & Gravano, L. (2015). K-shape: Efficient and accurate clustering
of time series. In \emph{Proceedings of the 2015 acm sigmod international
conference on management of data} (pp. 1855–1870). ACM.
}

\seealso{
Other cross-correlation functions: \code{\link{crossCorNormalized}}
}
