% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subVectorMean_fast}
\alias{subVectorMean_fast}
\title{Mean with Weighted First And Last Element}
\usage{
subVectorMean_fast(x, startIdx, endIdx)
}
\arguments{
\item{x}{Numeric vector.}

\item{startIdx}{Index of first element of sub-vector.}

\item{endIdx}{Index of last element of sub-vector.}
}
\value{
The weighted mean as double.
}
\description{
Calculates the mean of a sub-vector, weighting start and end if the corresponding
indices are no whole numbers.
}
