% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{tsCrossDistMat}
\alias{tsCrossDistMat}
\title{Compute cross-distance matrix}
\usage{
tsCrossDistMat(tsList, distMethod, distArgs = list(), distExports = NULL,
  distPackages = NULL, cpuCores = 2, maxSequentialTime = 3,
  trace = TRUE)
}
\arguments{
\item{tsList}{A list containing the objects (e.g. numeric vectors representing
time series).}

\item{distMethod}{The name of a function which should be used to calculate the
distance between two objects.}

\item{distArgs}{Further parameters which should be passed to the distance
function in the form of a named list.}

\item{distExports}{Function (not part of a package) which are need for the
distance computation and therefore need to be exported to work with
 \code{foreach}.}

\item{distPackages}{Packages which are necessary for using the \code{distMethod}.}

\item{cpuCores}{The number of cores/threads which should be used for
parallelization. Setting this value to 1 forces sequential computation.}

\item{maxSequentialTime}{A short speed test is performed to decide if
parallelization is worth it (as it comes with a certain overhead). If the
speed test expects a total runtime <= the parameter value (in seconds), it
calls the distance method sequentially. A parameter value <= 0 forces parallel
execution (without speed test).}

\item{trace}{Print status information? (currently only the computation time
after finishing the computation).}
}
\value{
A numeric matrix with row and column count == number of objects in
\code{tsList}, containing the pairwise distances.
}
\description{
Computes the cross-distance matrix between all (time series) objects from a
list, assuming a symmetric dissimilarity and a self-distance of 0. Uses
parallelization.
}
