% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vanDongen_fast}
\alias{vanDongen_fast}
\title{Van Dongen Criterion}
\usage{
vanDongen_fast(assignments1, assignments2, normalizeAndInvert = FALSE)
}
\arguments{
\item{assignments1}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters (code depends on this!).}

\item{assignments2}{Integer vector of cluster assignments containing only values
from 1 to k with k = number of clusters.}

\item{normalizeAndInvert}{Should the van Dongen criterion be normalized to
[0,1] and inverted such that high values indicate similar clusterings?}
}
\value{
The van Dongen criterion as double (in [0,2n) without normalization
and [0,1] else).
}
\description{
Calculates the index of van Dongen (2000) to compare two cluster assignment
vectors (external cluster validation). It is a value in [0,2n), lower values
indicating more similarity (it matches each cluster of one assignment to the
most similar cluster in the other assignment and counts mismatches). Optionally,
the index can be normalized to [0,1] as proposed by Wu, Xiong and Chen (2009).
After normalization, we take 1 - normalizedValue so that higher values indicate
better clustering quality (as it is for indices like Rand, Fowlkes-Mallows).
}
\section{References}{


Van Dongen, S. (2000). \emph{Performance criteria for graph clustering and markov
cluster experiments}. National Research Institute for Mathematics and Computer
Science. Amsterdam.

Wu, J., Xiong, H. & Chen, J. (2009). Adapting the right measures for k-means
clustering. In \emph{Proceedings of the 15th acm sigkdd international conference
on knowledge discovery and data mining} (pp. 877-886). ACM.
}

\seealso{
Other External Cluster Validity Indices: \code{\link{conditionalEntropy_fast}},
  \code{\link{fowlkesMallows_fast}},
  \code{\link{pairCVIParameters_fast}},
  \code{\link{phi_fast}}, \code{\link{purity_fast}},
  \code{\link{randIndex_fast}}
}
