% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorCrossDistMat}
\alias{vectorCrossDistMat}
\title{Pairwise Absolute Distance}
\usage{
vectorCrossDistMat(x, y)
}
\arguments{
\item{x}{1st numeric vector/time series.}

\item{y}{2nd numeric vector/time series.}
}
\value{
An |x|*|y| matrix containing the pairwise absolute distances.
}
\description{
Computes the pairwise absolute distance between two numeric vectors
(e.g. univariate time series)
}
