% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceComputations.R
\name{znormalize}
\alias{znormalize}
\title{Z-normalization}
\usage{
znormalize(x)
}
\arguments{
\item{x}{A numeric vector/matrix (uni- or multivariate time series).}
}
\value{
Z-normalized time series.
}
\description{
Calculates the z-normalization by subtracting the mean and dividing by the
standard deviation (constant series are normalized to 0). Multi-variate time
series (matrices) are normalized per column.
}
